Load app from dynamic assembly
This adds UnoAppLoader, which looks up information about the app in
Info.plist before loading and starting it.

Before, we (the Uno compiler) used to edit the app-loader assembly
using Mono.Cecil and insert the startup code for an app, but editing
assemblies based on Microsoft.macOS.Sdk on .NET 6.0 leads to crashes.

Now, we (the Uno compiler) can insert the right information in
Info.plist instead of editing assemblies on macOS.